refactor(all): re-set numeric types

+++ src/main/java/com/lian/marketing/lianstockmicroservice/domain/model/Product.java
@@ -15,8 +15,8 @@ public class Product {
     private UUID id;
     private String name;
     private String description;
-    private Double stock;
-    private float priceSell;
-    private float priceBuy;
+    private Integer stock;
+    private Double priceSell;
+    private Double priceBuy;
     private Subcategory subcategory;
 }
